Remove niche optimization assertion to make the generated code slightly less crufty.
I think this will implicitly be caught by layout assertions, and anyway is just a core part of the language -- we could also, equally validly, check if isize and pointers are the same size, and so on.

In the interest of making it easier to read the generated source code (for documentation, etc.), I want to strip it out. As a side benefit, one assumes everything will be ever so slightly more efficient.

PiperOrigin-RevId: 602650817
Change-Id: I3876648da8c3edb3d670be70a6b2f57d81eb4eaa