Use the enum example as a source of truth for the reference documentation.
This uses a slightly nasty regexp to filter out all comments and crate level attributes from the golden output file, giving you just the bindings themselves.

This is, more or less, a proof-of-concept -- if we like how this works (ugly regexp and all), we can do all our docs this way. Otherwise, we may want to filter these from the golden itself, via e.g. a flag to Crubit and a transition. (A bit difficult/annoying...)

PiperOrigin-RevId: 602970122
Change-Id: I6e8c59ff45a389181a11567b97f286e68557fd2f