[nullability] Add a convergence test based on Abseil's `LOG_FIRST_N`.
This expands to three nested for loops and led to "max blocks visited" errors
when ymand@ initially tried to change the order in which block successors are
visited in the dataflow framework
([#80030](https://github.com/llvm/llvm-project/pull/80030)).

We now no longer see these "max blocks visited" errors with this test because
of
[other](https://github.com/google/crubit/commit/3dfd35e2a98339957a1087f53a9788b1dd342a14)
[fixes](https://github.com/llvm/llvm-project/pull/80033) that have landed in the
meantime. Nevertheless, it seems worthwhile to put this test in place to prevent
regressions.

PiperOrigin-RevId: 603020024
Change-Id: I7a8428a77a2f4252d012eec30e42ccb0c1b96bc5