[nullability] Fix early-outs in custom `merge()` operation.
Our intent is to signal "perform the default behavior", i.e. keep the
`MergedVal` that has been produced by the framework. This would be done by
returning true (which is also what the [default
implementation](https://github.com/llvm/llvm-project/blob/d5c9d402f07e7448cd46870a59d981f582682548/clang/include/clang/Analysis/FlowSensitive/DataflowEnvironment.h#L105)
does).

However, we were returning false, which means that the framework should drop
the `MergedVal` and remove the corresponding entry from `LocToVal`.

This was initially causing "max blocks visited" errors on the
`TriplyNestedForLoopSingleIteration` test when ymand@ tried to
change the order in which the dataflow framework visits block successors
([#80030](https://github.com/llvm/llvm-project/pull/80030)).

We now no longer see these errors because of
[other](https://github.com/google/crubit/commit/3dfd35e2a98339957a1087f53a9788b1dd342a14)
[fixes](https://github.com/llvm/llvm-project/pull/80033) that have landed in the
meantime. Nevertheless, it seems worthwhile to fix `merge()`, as this reduces
the number of block visits required to achieve convergence. On the
`TriplyNestedForLoopSingleIteration` test, we require 66 block visits before
this patch, and 53 block visits after.

The issue was that when performing the join at the loop head, our custom merge
logic was causing the merged value for the variable `int c` to be discarded from
`LocToVal`. This in turn was blocking convergence because widening
[considers](https://github.com/llvm/llvm-project/blob/d5c9d402f07e7448cd46870a59d981f582682548/clang/lib/Analysis/FlowSensitive/DataflowEnvironment.cpp#L626)
two environments to be different if the sizes of their `LocToVal` maps differ.

As an aside, widening is inconsistent in this respect with the convergence check
that we perform on blocks that don't get widened.
[`Environment::equivalentTo()`](https://github.com/llvm/llvm-project/blob/d5c9d402f07e7448cd46870a59d981f582682548/clang/lib/Analysis/FlowSensitive/DataflowEnvironment.cpp#L570),
which performs this check, uses `compareKeyToValueMaps()` to compare the
`LocToVal` maps of the two environments, and `compareKeyToValueMaps()` only
compares values for locations that are in both `LocToVal` maps and ignores
locations that only occur in one of the maps. We should probably fix this
inconsistency.

PiperOrigin-RevId: 603025629
Change-Id: Id02993e576023c44a335c8991020f7e233a75320