rs_bindings_from_cc: Don't invoke `coverage_common.instrumented_files_info` when running `rustc_compile_action` in aspect.
To do that, we also need a patch to rustc.bzl so that we can toggle whether or not `rustc_compile_action` invokes `coverage_common.instrumented_files_info` to create `InstrumentedFilesInfo`.

Why? Given that a cc_library with name `a` that is used as a `cc_deps` in a rust binary, when computing the coverage, Bazel

- Computes the coverage of the cc_library itself, creating the coverage file `a/baseline_coverage.dat`, and

- Computes the coverage of the Rust crate that has the Rust bindings (which may be empty if Crubit is disabled) of the cc_library when the rs_bindings_from_cc_aspect is propagated, creating `a/baseline_coverage.dat`.

This causes `a/baseline_coverage.dat` to be generated by 2 different actions, and is thus considered as a conflict. Disable coverage for generated bindings for now, because the build failure blocks coverage collection for the Rust crates.

(`cc_bindings_from_rs` direction of the interop doesn't have this issue because the `cc_bindings_from_rs_aspect` is attached to the custom rule `cc_bindings_from_rust`, which has a different name than the rust_library.)

PiperOrigin-RevId: 603072316
Change-Id: I176a7b88ef209191780545a50533cf4837eb14c0