[clang][nullability] Treat classes derived from smart pointers as smart pointers.
Tensorflow's
[`RefCountPtr`](https://github.com/tensorflow/tensorflow/blob/932d770cc98f26ff2aa372b08a020093bc718d25/third_party/xla/third_party/tsl/tsl/platform/refcount.h#L101)
does this, for example.

It's not an option to simply ignore these derived classes because they get cast
back to the smart pointer base classes (for example, simply when calling the
smart pointer member functions), and our transfer functions will then run on
those base classes and therefore require them to be properly initialized.

Treating this case is complicated by the fact that the
`ClassTemplateSpecializationDecl` for the derived class may not be instantiated,
in which case it doesn't even have a `CXXBaseSpecifier`; see also the comments
in the code.

PiperOrigin-RevId: 603348664
Change-Id: I843b9488858ef8b813b7fcca337d763b45ac7aac