README.md revamp.
In broad strokes, the goal of this CL is to convert the README.md into a good "one pager" description of Crubit as a whole: what does it aim to be, and where is it today? Before this change, the README.md did not lay out the main distinguishing goal of Crubit, the thing that makes it different from existing tools like cxx or bindgen: high-fidelity interop. Of course, once we do that, we need to also be very clear about the difference between the "end state" and the "current state", which I've tried to do here.

---

I've also, in this CL, converted the external-do-not-use warning to something _specific_. That is, instead of FUD-ing ourselves and making a vague "don't use this" claim, I've tried to lay out the specific reasons why an external user should not use us.

PiperOrigin-RevId: 603602753
Change-Id: I3efc9014853bd64259df840b443bc789db9a5521