Deduplicate unknown_attr checks.
These were repeated (for enums and type aliases) across rs_snippet and src_code_gen, basically just because of personal confusion / I forgot why we had some logic in each file.

This kind of check only needs to happen once, when the RsTypeKind is initially created / in has_bindings, which means it can be removed from the checks in rs_snippet.

PiperOrigin-RevId: 603606753
Change-Id: I834b8b384623973679bf173a7be1587d3c90a995