Omit bindings for namespaces with unknown attributes.
This _should_ also include everything _inside_ the namespace, since we have no way of exporting them if the namespace itself is gone! But that will wait to a followup CL for organizational reasons -- indeed, we already have the same bug for structs (if a struct is gone, its nested types aren't). That's a problem with has_bindings rather than with required_crubit_features.

PiperOrigin-RevId: 603607422
Change-Id: I0fb30420b7e07c41d7a0c8176d7e26bc6182e72f