Unify parent namespace / records, and reject types nested in records.
Where items like enums were ignoring the parent record and being placed in the parent namespace instead, this now (correctly, like with type aliases) hides the item entirely instead. I was forced into doing that fix in this CL, because, once the two are merged into one field, the parent record _hides_ the parent namespace.

The intended longer-term effects are:

1) everything which had a parent namespace now can also report a parent record. There's no more careful patching of it -- it's supported across the board. Because the record name now _hides_ the namespace name, bugs like the above are forced to be fixed.

2) we can now write logic which is agnostic about what kind of parent it is. For example, the logic I wanted to write, which spurred this CL, was: "if the parent item does not receive bindings, then neither should this". But in fact, it wasn't possible to write that logic in general, we had to write it once for every item type, and once for record parents and once for namespace parents. So, that was the motivation for this. (The CL I actually want to write is now just an extra line near the nested-item check, to check has_bindings(parent).)

Anyway, at a high level, if the direct parent is a struct, we don't need to worry about the namespace ancestor (except through the struct), and if the direct parent is a namespace, we don't need to worry about if there is a struct parent. So having two parent IDs didn't make sense.

PiperOrigin-RevId: 603614209
Change-Id: I50fa1d5bbf5d90ba36e65a207983738e38401659