Universalize the type alias passthrough workaround.
With this change, if the bindings for the type alias are disabled, we see through the type (if we can), across the board.

The only exceptions are if it has an unknown attribute, or is in a target that doesn't support Rust. These are explained in more detail in the comments.

... actually, those exceptions mean that this change is a no-op!

**However**, it will not be a no-op if we add new reasons to disable bindings. Indeed, I intend to add one in a followup CL, which is similar to the nested type case: if the parent type is unsupported, then typedefs should still "see through". Refactoring the logic in this way makes that followup CL a fair bit easier, because the decision happens all in one place instead of distributed across the bindings generation.

PiperOrigin-RevId: 603618082
Change-Id: Ib73b123a3abd22d1c5344f832aad0b7d5f6a69a8