Collect Nullable evidence from assignments from null/nullable.
E.g. `p = nullptr` indicates that p should be Nullable.

This notably diverges from the way we treat Unknown-but-provably null values in other places in evidence collection, but in a more thorough way. Left TODOs to perhaps adjust those other places to be more like this, and added/adjusted tests to make current behaviors clear.

PiperOrigin-RevId: 603704781
Change-Id: I193b23454af0e7bf2c015b29800cb51d126bb1ed