Per discussion with Sam McCall, drop the unknown attribute check.
We use the Clang AST! We don't need to resort to that kind of syntactic worry. The underlying type / aliased type of a type alias is _always_ correct in the Clang AST, and we don't need to care what the attributes do -- to the extent that they change what the alias does, that will already be reflected.

And so it is safe to look through even a type alias with an unknown attribute. Or, well, as safe as anything is with unknown attributes.

In the specific case of the vector attributes, they'll fail to compile because we don't have any support for the builtin vector types -- they need to be hardcoded the same as integers, so far as I know. (I didn't actually check the output, because I don't have command-line access atm, because Reasons.) But whatever happens, it certainly won't accidentally become just a regular `int`, which was the worry.

PiperOrigin-RevId: 604168828
Change-Id: Ia1034ecfd68fbccd2060c418a480f90f3edfb83f