Deny bindings for types with unknown attributes.
This should, AFAIK, be the last change for denying unknown attributes in extern_c for rs_bindings_from_cc. We do it now for all top-level items, for parameters and fields, and for types.

This is distinct from parameters/fields/items in that this restriction also applies in `:experimental`. This is... ~fine. There's almost no type attributes anyway (the only major one left is the nullability attributes, but then that's basically it), and it isn't important to "keep going" in :experimental anyway.

The reason this one is handled differently is that it's a challenge to try to keep the context of what mode the type is being translated in, as well as a drain on performance because it reduces cachability. Being less convenient, slower, and unimportant, it seems like a straightforward decision: let's not bother trying to check whether the type is being used in an extern_c context.

PiperOrigin-RevId: 604242901
Change-Id: Ie40cbeab915eef4d13fb4d2969b0da65f5bd94c7