Rough cut: split reference docs into `/cpp/` and `/rust/`.
This hasn't gone through an extensive editing pass, but is just the basic split for the most part.

In this split, `/cpp/` contains information about C++ APIs -- how does a C++ API map to Rust? `/rust/` contains information about how Rust APIs project to C++.

The first-order motivation for this is that the C++ half is going to be ready to use significantly earlier than the Rust half, and so it's difficult to try to keep the docs merged, as it means the docs are ~50% not relevant to the reader at all.

Maybe we should expect this to always be true? Like, your task is either to call a Rust API or to call a C++ API, not to call a both API. I think Lukasz made the same point when these docs were written.

But at a high level I still feel that bidirectionality is something significant in the docs -- structured right, surely the `/cpp/` docs _are_ the bidirectional docs, because every C++ interface should be able to be called from Rust (rs_bindings_from_cc) or implemented in Rust (cc_bindings_from_rs). If it's the same code either way -- if crubit(crubit(c++)) == c++ -- then the docs for implementing a C++ interface in Rust or using it in Rust are the same thing, it's just whether you read them from left to right or right to left. We _aren't_ there yet -- for instance, enums are certainly broken in this regard -- but it is the long-term goal so far as I know.

So in theory, maybe the `/cpp/` docs are inherently the bidirectional docs, and `/rust/` docs are the leftover Rust features that are not just a C++ API projected into Rust, but novel and unique to Rust. (Crubit is inherently a little asymmetric in this way: it's important that Rust can do anything C++ can do, but we have no converse requirement.)

---

In that vein, the following extra changes, not just splitting, have been performed:

* The Rust enum docs are merged into one file (`enums.md`). The old docs made the distinction between fieldless and fieldful enums, but this is, I think, the wrong distinction -- we will draw a distinction between C++-like enums (using a special proc macro) and non-C++-like enums, but all Rust enums are alike, because none of them can be mapped to a C++ enum.

---

P.S. I should reiterate, the main goal here is to be approachable when only one half is released. So I think it's fine if we don't yet have the broader picture fully fleshed out, as long as we're moving towards that shorter term goal. We probably cannot jump straight to the ideal final state, anyway.

PiperOrigin-RevId: 604685146
Change-Id: I3eb9d5f22817a0e452f5915bcf1debf71367c86f