cc_std: Move transmute-based slice_ptr_get into its own file.
Still, slice_ptr.rs needs to be in `cc_std` target, because, currently, we don't have the starlark infrastructure to have `cc_std` depend on a Rust crate (everything is included as a `filegroup`...).

But splitting it out into its own file is the step in making it easier to share this utility with other Rust targets.

PiperOrigin-RevId: 604907385
Change-Id: Id21cab2b4987aa5183b9192d531d891fbe226b98