Small cleanup: don't add `C`'s preconditions to `~C`.
The destructor doesn't need to include all feature requirements for the class it is a destructor for. The requirement that the class itself be supported is already implicit in has_bindings -- so all this feature check serves to do is make the true cause of the error less clear (by converting it from a dependency error to a direct error).

This is already implicitly tested by whether or not the bindings for `no_bindings.h` (which has a class and destructor) compile.

(The main motivation for this cleanup is to nix the recursive call, tbh, since it makes it easier to handle error reporting and fix b/318006909.)

PiperOrigin-RevId: 605191864
Change-Id: I0097c971300dcd592d6e9c54d7557cbc63dbe6e6