Give a **reason** that something requires a given feature.
The resulting output is a little redundant -- maybe we can work on that -- but finally actually points to specific things. For example, if `Foo` is a non-extern "C" function, it will say something like:

```
// Error while generating bindings for item 'Foo':
// Can't generate bindings for Foo, because of missing required features []
// //my:target needs [//third_party/crubit/features:experimental] for Foo (non-extern "C" function)
```

This isn't quite enough to solve b/318006909, because this only explains the reason for non-type items. For types, we need another similar refactor, in a followup CL.

PiperOrigin-RevId: 605235688
Change-Id: I5fe5e560f0f08b6553833d209fbef9655339177c