rs_bindings_from_cc: Only import methods of class template specialization that can be instantiated without 'error' diagnostics.
When a method of a class template specialization is imported, Crubit generates a thunk that invokes the method, which causes it to be instantiated when compiling the generated bindings. If this causes an error level diagnostics to be emitted, then the generated bindings will fail to compile, failing the entire build.

To avoid this problem, we first try to instantiate the method and examine the diagnostics. (Alternatively, I suspect we could generate thunks in a way that don't cause the method to be instantiated if not invoked in C++, which should give us guarantee the method can be instantiated, but I'm not familiar with C++ templates to know how...)

There are 2 caveats to always eagerly instantiate the method of a class template specialization:

- It does not instantiate the templates in the function template body, i.e., not recursively, which means we won't be able to catch diagnostics for them, which means there still a chance that we import an un-instantiable method :) More in the comments and test.

- It mutates the AST, after its creation. I don't know how it impacts other parts of the logic, but it seems fine for now, as I didn't need to update any existing tests.

PiperOrigin-RevId: 605569636
Change-Id: If86bd6b39f11c01477c3074dc0552262d1c5b131