nullability_test: Fix TEST on constructors, clean up nullability_test_tests
This was supposed to work but hasn't been used yet, so we have silly assumptions
that functions have names.

While adding tests: clean (a bit) the horror show that is nullability_test_test.
These were supposed to be smoke tests, but it's clear that's not how they're
actually used, so we should avoid the boilerplate.

PiperOrigin-RevId: 606597645
Change-Id: I83f466777548bf206d6b65e7c5b572409b462801