Move `char32_t` etc. support into `ConvertUnattributedType`.
This doesn't change any behavior, it's just the right place for it. These are builtin types, so we don't need to fall back to the special casing for library types.

(Well, I would guess it's slightly faster.)

This logic is tested already in types_test.rs, and it's ~fine: there are no invariants for the values, unlike Rust `char` etc.

PiperOrigin-RevId: 606624387
Change-Id: I7f6e78d1bff6f75f0eb8ff8275b706a71cd526de