Allow arguments to be passed in an param file.
This resolves the issue of "too many arguments" in general. Actually, I'd _hoped_ it'd only be for positional args, but our toolchain passes _all_ arguments through here, which is why we need to unpack it eagerly and process the flags and so on.

By doing the logic inside the (public) rs_bindings_from_cc, instead of the (internal) glue driver bash script to the toolchain, this solution is also reusable outside of the monorepo for things like chromium.

(Also, it means I have less bash to write.)

---

It is _tempting_ to use `--flagfile` -- and I initially argued as much -- but `--flagfile` doesn't support positional arguments, and translating between positional and flags is not at all trivial and would require, well, all this glue anyway.

This CL (and its predecessor) is basically the same as the previous experiment by Jing Lu (unknown commit, unknown commit, and unknown commit), but it merges the extra binary into rs_bindings_from_cc, as C++.

PiperOrigin-RevId: 606979458
Change-Id: Ie00070d4a5035f4afe91fca4c9ead6fb9b2da06e