Also add an example of using `clang::trivial_abi` to define a struct.
The generated output is not the most readable, we may be able to improve it a bunch. For example, some of the asserts are silly, and it's generating instantiations here for the header `attributes.h` which isn't ideal. However, those are probably annoying enough that it'll have to wait until later.

PiperOrigin-RevId: 606981997
Change-Id: I3e0c486feb44aa3184110dde7452f3a2c4b7aab4