Only support forward declarations for structs with :experimental.
In all other feature sets, this suppresses the definition marker, meaning that a forward declaration cannot be `cc_cast` cast to the real type.

Not really a big deal, since who is going to use forward_declare anyway? But it helps close off an unintended feature, and simplifies the generated output.

PiperOrigin-RevId: 609058288
Change-Id: Ie457847d9af125b8b1e77b88e1ad04a53ec524f2