Write safety comments for string_view helpers.
(Also, restructure the code to use finer-grained unsafe blocks so it's clear what each safety comment applies to.)

Of course, this uncovered an obscure issue in `From<string_view> for *const [u8]`. I wrote this function before I became familiar with the exact details of dangling pointer semantics in Clang, which... are really confusing, and not what you'd expect from reading the standard. (To be honest, I don't really accept in my heart of hearts that the standard _permits_ the optimizations Clang performs -- it treats "implementation defined" the same as it would if it were "undefined", but if it were OK to do that, we wouldn't even need "undefined behavior" as a category: everything could be implementation-defined or _unspecified_ behavior. That is: I had always thought implementation-defined behavior was supposed to _have_ a behavior, and not allow arbitrarily bad things. But, here we are. I am not a language lawyer!)

That issue is not fundamental to the _API_, only to the _implementation_, and so I am not entirely dissatisfied with leaving it as a todo. (In fact, we might be very close already to `size()` working -- I will try it when I get SSH access again.)

I have noted the increased urgency of this in the attached bug, b/249376862.

PiperOrigin-RevId: 609139779
Change-Id: I076713261d1869dccfd08a264300cc11f8dc10df