Add Crubit docs for class types.
This is, of course, a bit tricky to document well, but I've done my best. Feedback / suggested edits welcome.

The example is a bit hairy because of the assertions. I think we can fix this, but I can't do it until next week, when I get SSH access back and can regenerate goldens and stuff.

PiperOrigin-RevId: 609160728
Change-Id: Ia99b648ad23e1c0e34d0e455f4ffafb5c0b1be33