Update/revamp function pointers doc.
Most of the supplementary information is moved to footnotes, some is deleted.

I also expanded out the concept of a "thunk" -- the user probably does not know what a thunk is, and we do not document when one is generated. Instead, what matters is the notion that not all function pointers are possible to call from Rust at all. So I document, here, when that is, and what the workaround is.

PiperOrigin-RevId: 609161808
Change-Id: I68c9a6ca960db779ce59b053c0b4adf8bfeca593