[nullability] Crash fix for modelling of non-const methods.
These could produce pointer values without nullability properties; see the crash
repro included in this patch.

To fix this issue, `handleNonConstMemberCall()`, instead of producing a
`PointerValue` without nullability properties, now clears the value of the field
entirely. In this way, we continue to uphold the invariant that all
`PointerValue`s have nullability properties.

PiperOrigin-RevId: 610660134
Change-Id: I12ecf8bb77bb40f3b28d7b2d857d29bb64859763