Group all assertions into one `const _: ()`.
This makes the result slightly easier to read, by deduplicating the boilerplate necessary to run the assertions. It also removes some jarring inconsistencies with how different types of assertions were written out. (Some used an extra `{}`, some did not -- it's this part that results in a net LOC reduction in the output.)

PiperOrigin-RevId: 610848343
Change-Id: I64002071a069b1db99cd30d20f4f53d26941fd27