Create a "broken" example, for use in a codelab/introduction.
This allows users to actually run crubit on the broken target and see what is broken about it, and cite it in introductory material.

(I had considered just pointing them to run crubit at //third_party/absl/time, but the transition to "and now time to make a working library" doesn't work well there.)

It's desirable to split this to a separate CL from the documentation using it, so that the preview works right -- need code to be _submitted_ for it to work in live snippets.

PiperOrigin-RevId: 611580637
Change-Id: Id5be7b8f371d5047b3440114fabeaf68d6b3303f