[nullability] Reorganize functions for getting pointer values from expressions.
While working on a crash, I realized that `getPointerValueFromExpr()` has a
potentially misleading name in that it doesn't make it clear that it only works
for raw pointers.

This patch improves the names and additionally introduces a new function that
can get a pointer value from a raw _or_ smart pointer expression.

*  Rename `getPointerValueFromExpr()` to `getRawPointerValue()`. (I think the
   `FromExpr` is clear from context, so it's redundant.)

*  Rename `getPointerValueFromSmartPointerExpr()` to `getSmartPointerValue()`

*  Introduce `getPointerValue()`, which can retrieve a pointer value from a raw
   or smart pointer expression, and use it in the two places where we had so far
   been making this case distinction locally.

PiperOrigin-RevId: 611729790
Change-Id: Id9132a2b53142632e344dff2d35622f69a9f122e