Fix UB in no_unique_address_fields.
Ironically enough, we already had a UB-free implementation for empty fields (which occupy _no_ bytes of the object) -- we just didn't use it unconditionally. We had to do the right thing for those because the field wasn't present on the Rust type definition at all, and so there was no field to reuse for a bad pointer cast.

Unfortunately, the UB-free implementation for empty fields had a bug in it, meaning it had UB after all, giving you pointers way off the end of the object. (It used bitsize where it should've used bytesize.) Buuuuuuuuut I fixed that.

This CL cleans up the code, fixes UB, and even adds a safety comment. hooray! hooray! hooray!

PiperOrigin-RevId: 613008832
Change-Id: Idd1bf7641bcb19a24012241e76488f60a50936e6