Explain why types require a given feature.
This was a bit messy, I am not entirely sure if it's possible to make it nice and clean. Maybe I need to take a look at it with totally fresh eyes.

In the interest of not stating the obvious, I only include an error message if the requirement is not for ExternC. If you don't have crubit enabled at all, you already know the deal. For instance:

* non-Unpin types (the reader might not realize it was not rust-movable)
* Incomplete types (the reader might not realize it was incomplete)
* non-`extern "C"` function pointers (while you can determine this by reading, it's subtle)

PiperOrigin-RevId: 616098287
Change-Id: I331a827174babd57d729b5159776a8b1212d7151