Add a `--config=crubit-genfiles` to inspect generated C++/Rust bindings.
Note that, unlike most `--config`s, this is not actually useful for building binaries or anything like that. This is just a global shortcut for the aspect and its outputs. So it is primarily useful with `build` (for end users who want to look at bindings), and with `aquery` (for developers).

(I think there's no easy way to visit the aspect using `query` or `cquery` or the like, certainly not just with the `--aspects` flag alone.)

PiperOrigin-RevId: 617391700
Change-Id: I2df3964a6f5444c0089e8bc2af9d63ce76c08779