Remove `feature(more_qualified_paths)` from tests.
The long-term solution should be a type alias, not a proc macro. In the short term, we aren't releasing this feature yet. If we did release it, and we could not figure out how to make it a proper type alias, _and_ we could not stabilize the feature, the workaround is probably fine, anyway.

PiperOrigin-RevId: 620020591
Change-Id: I3e92cfb3897ca41107893e11eb971cbcb91a9810