Thread through split debug info for cc_bindings_from_rs.
This was a lot trickier than the previous CL. Before this change, the code being modified here tracked all the dependencies and added them to the linking context, but otherwise did not keep them around -- it did not keep the full CcInfos list to use for compilation_context or debug_context. This meant we lost the debug contexts, and also meant that even if we changed this into a list of CcInfo objects, we were missing a couple (specifically, the ones for the target itself, and for the generated Rust helper code.)

We apparently need the full list for debug_context, most of the list for linking_context, and very little of the list for compilation_context? I don't know how this works. But it is simplest and most expedient to use the full list for all three, and so far as I know this produces no ill effects -- and if it changes any behavior, one suspects it's a bug fix.

In fact, to prove the point, I specifically changed the logic here to mirror rs_bindings_from_cc 1:1: rs_bindings_from_cc *does* use the full list of all the CcInfo providers for all three, and does so by pre-emptively merging them using merge_cc_infos. I've done the same now, here, so that rs_bindings_from_cc and cc_bindings_from_rs produce the CcInfo in almost exactly same way.

PiperOrigin-RevId: 621348660
Change-Id: Ia45f26285b9ec25115ce2e4ccdcba643d18b4464