Basic support for nullability #pragma in diagnosis & clang-tidy check
This covers the diagnoser's existing use of the shared analysis, but does not
cover the places where the diagnoser has its own logic to determine the
nullability of "sinks" that expressions should be checked against.
test/pragma_diagnosis.cc documents the current sad state.

Converted checkDiagnostics to use TestAST so we can install the pragma handler.

PiperOrigin-RevId: 621491603
Change-Id: I62d46d8cd31d95a0cbaa9a9fb17a5007999f7ffb