Make use of type_nullability for LHS in diagnoser
Diagnoser checks fit one of two patterns:
- deref-like: *RHS. (operators *, ->, [] etc)
- assignment-like: LHS = RHS (return, pass to function param, etc)

For assignment-like checks, we switch to reading the nullability using `getTypeNullability()` instead of reading from the AST directly. This means:

 - we benefit from resugaring of the LHS type
 - we correctly interpret the #pragma governing the LHS type

This makes the pragma work usefully for the verification tidy check.
It also fixes a few scattered false-negative test cases.

Finally, it centralizes all the `LHS=RHS` verifications in one place, where it would be convenient to add checks for non-top-level pointers if we want that, using nullability vectors + variance information from the type to check compatibility.

PiperOrigin-RevId: 621548328
Change-Id: I932b93652e83a9f9fdb9c4d25111cb4b691229d6