Improve type-nullability precision by capturing template arg sugar in more places
The technique in substituteNullabilityAnnotationsIn* was sound, but it wasn't
applied in enough places: function calls would get only function template args,
member variable refs would get only base type args, plain declrefs would get
nothing.

Here we generalize this into a builder that supports two types of binding
(bind template param lists to a nullability vector, or to an argument list).
We call the builder in more places.
And we resolve an old TODO so that CallExpr doesn't have to worry about this
at all, leaving resolution up to the callee.

This fixes a bunch of longstanding false-negatives in diagnosis.

PiperOrigin-RevId: 621595924
Change-Id: I5dd25c10150b82e7a66464e222457d3a3410b406