Stop mangling names for `extern "C"` functions.
This fixes b/332718844 - we could not call non-`inline` functions which had `C` linkage, because we'd assume their name was mangled.

(This was not caught by tests because we often just used inline functions for tests, out of expediency. And for non-inline functions, we had it tested, but not specifically with `extern "C"`. Oops!)

PiperOrigin-RevId: 621627733
Change-Id: I1158051b3635fa9540d9064a144902dbf8977846