Add integration test covering all the basic types, as return values and fields.
This is analogous to rs_bindings_from_cc/test/types/types_test.rs, but less... Why did I do it that way? Why?

It does not include the `ffi` type aliases, because I intend to change them.

---

We do also have unit tests in bindings.rs, but they actually reproduce a lot of the logic of bindings generation, to the point that I'm not sure they give the same kind of assurance as an integration test like this. (For example, they need to manually call the functions for retrieving the rustc_middle ty -- and in one proposed CL, rustc_hir). Especially for structs, where there's actually quite a _lot_ of logic, this is both burdensome and introduces the possibility for divergence / uncaught bugs.

PiperOrigin-RevId: 623329457
Change-Id: I65ba5c7d9c8a2cb636ce620c6de1119b01a8491b