Fix `CARGO_MANIFEST_DIR`.
This allows e.g. pest to be used in transitive deps of crubit bindings.

This is a bit of a hack: we are still hardcoding a substitution provided by `--subst` in the rustc process wrapper. And we're hardcoding the list of environment variables it might apply to.

A more durable fix would be to read the process wrapper arguments and reproduce them and their semantics exactly, or else to run crubit via the process wrapper. The latter is probably best, and is the subject of b/277376088

PiperOrigin-RevId: 623954130
Change-Id: I0f4087eb1c0a7814eca2ed8a16e075f5274b0901