[nullability] Fix crash in `initializeOutputParameter()`.
This would crash if the parameter was a pointer-to-member to a pointer (see
newly added test).  Such a parameter would get past the test
`ParamTy->getPointeeType().isNull()`, but `getTypeNullability()` would return an
empty `TypeNullability`.

It makes no sense to handle pointers-to-members here anyway (we don't know what
they point to without having a corresponding object), so just bail out if we
see them. (This is likely what we intended anyway.)

PiperOrigin-RevId: 625601825
Change-Id: Ieafff4121167aa35268fb75d1187cf4a15b27d57