Unconditionally add `clang::trivial_abi` to a type's attributes.
This has the following effects:

1. (primary benefit) it is easier to use these types from manual FFI -- they can be passed directly by value, not by pointer.
2. the types are faster to pass by value in pure-C++ code.

This does _not_ improve performance passing the types by value to Rust, as we still use a thunk that takes them by pointer -- just in case they have any fields which were replaced by an opaque blob of bytes.

Since there's no effect on automatically generated FFI, this might be the first time we have manual FFI in a crubit integration test.

### Correctness

This is always correct: if a Rust type exists by _value_, then it is safe to pass via registers.

(In fact, if it is `repr(C)`, it _must_ be passed via registers when FFIing to a native Rust functioin. For non-repr(C) types passed by value across FFI, we need to use a thunk that passes by pointer, and the `trivial_abi` is irrelevant for FFI, and merely an optimization for pure-C++ use. In any case, we currently thunk all of them and use a pointer, so it's a moot point.)

We will eventually want to support creating non-`trivial_abi` structs from Rust, but there's no support for this at all yet, so it's fine to be unconditional.

### Testing

As expected, before the attribute is added, the newly-added test fails. Strictly speaking, what we got was UB, but at runtime we get a garbage result:

```
assertion `left == right` failed
  left: 12.34
 right: 6.9529741790386e-310
fatal runtime error: failed to initiate panic, error 5
```

After the attribute is added, the test passes.

PiperOrigin-RevId: 625632694
Change-Id: I8bbfce8f8eb4681ec2bcb8127001114d4f1eb6ee