Use a stable output for printing type alias errors.
When we fail to print the type alias, we print the debug string. This includes a numerical id field which appears to be unstable. Because of this, the build step is non-deterministic.

Instead of printing the full debug string, implement a display string that just prints the name/source info of the type alias. The other stuff may be useful for debugging, but likely not needed for the generated file.

PiperOrigin-RevId: 626445018
Change-Id: Ica8b1deb193788fbf1399c4e44d732e52badb2ec