Update Abseil version using bazelmod.
Abseil now [only supports bzlmod](https://abseil.io/docs/cpp/quickstart#set-up-a-bazel-workspace-to-work-with-abseil), and bzlmod is also the primary not-legacy workspace option in bazel itself. IDK, I guess we need to get with the times.

This change, as part of updating abseil, enables the hybrid mode, using bzlmod only for the abseil dep because the abseil dep is what was out of date: we need no_destructor. See https://bazel.build/external/migration

The remaining entries in WORKSPACE.bzlmod can be migrated as needed I suppose, or when we're forced to by deletion of legacy functionality or what have you.

PiperOrigin-RevId: 626477305
Change-Id: Ic34384b6fa37323aeca1d1e127f7dd81297b8ce5