Use C++20 in Bazel CI.
We use C++20 internally now, and so does Chromium for that matter. It's 2024, it's safe now.

(For example, we use `emplace_back(aggregate, members)` -- direct initialization of an aggregate -- which is C++20. Also, `using enum`, and who knows what else. We could fix these individually, but it is simpler and better to move right to C++20.)

PiperOrigin-RevId: 626483163
Change-Id: I8cf8cef926f842e96746ff658c24c60b6895e89c