Avoid C++20 direct-initialization aggregates.
This feature was only implemented in Clang 16, which is too new to be in CI. :'(

(Once we update to Ubuntu 24.04 for CI, it'll be fine, I think.)

Once this CL is in, Crubit actually builds in CI, it just doesn't pass tests entirely. Really close to the finish line, now.

PiperOrigin-RevId: 627133336
Change-Id: I8df710635c46707b37af266c7eab5cb49ce0cbb0